[YOLOv11] CI fixes, CodeQL alerts, packaging, and PR-readiness cleanup
Squash 19 follow-up commits (837dff31..52200486) that addressed CI
failures, CodeQL alerts, packaging gaps, clang-format/doxygen issues,
and stale doc/comment references:

- Address Codex review: q8act scratch per batch, depthwise filter dtype
- Replace structured binding with plain locals for clang c++20
- CausalLM conformance: Factory-registered config-driven model + Tensor-based c2psa
- cpp quantizer: --conv_dtype graph-walk + Q8_0 save with repack_q8_0
- Retire non-standard tooling: drop python quantizer, env-wrapper scripts, custom docs
- Add missing doxygen tags to yolo_postprocess.h for CI static checks
- Declare Q8_0 indirect-conv gemm symbols for x86 to fix CI build
- Add x86 stub for nntr_gemm_q8_0_q8_0_4x4_fp16 to fix FP16 link
- Package q8_0_tensor.h in Tizen devel spec and Debian devel install list
- Widen fast-path index arithmetic to size_t (CodeQL high alerts)
- Widen slice fast-path memcpy size to size_t (remaining CodeQL alerts)
- Co-locate converter at res/YOLOv11/weight_converter.py per standard
- Rewrap weight_converter comment for clang-format-14
- Retire standalone jni/ harness; colocate layers at model root
- Fix stale docstring refs to removed harness/scripts
- Generalize the SiLU fast-path env var and neutralize model names
- Neutralize a model name in the memory-planner comment
- Correct stale doc comments: weight repack is C++, not Python

+++ nntrainer/layers/slice_layer.cpp
@@ -128,21 +128,22 @@ static void sliceForwardT(const Tensor &input, Tensor &output,
     } else if (axis == 1) {
       for (unsigned int b = 0; b < B; ++b) {
         for (unsigned int hw = 0; hw < Ho * Wo; ++hw) {
-          std::memcpy(out + (b * Ho * Wo + hw) * Co,
-                      in + (b * Hi * Wi + hw) * Ci + start, Co * elt);
+          std::memcpy(out + ((size_t)b * Ho * Wo + hw) * Co,
+                      in + ((size_t)b * Hi * Wi + hw) * Ci + start, Co * elt);
         }
       }
     } else if (axis == 2) {
       for (unsigned int b = 0; b < B; ++b) {
-        std::memcpy(out + b * Ho * Wo * Co,
-                    in + b * Hi * Wi * Ci + start * Wi * Ci,
-                    Ho * Wi * Co * elt);
+        std::memcpy(out + (size_t)b * Ho * Wo * Co,
+                    in + (size_t)b * Hi * Wi * Ci + (size_t)start * Wi * Ci,
+                    (size_t)Ho * Wi * Co * elt);
       }
     } else if (axis == 3) {
       for (unsigned int b = 0; b < B; ++b) {
         for (unsigned int h = 0; h < Ho; ++h) {
-          std::memcpy(out + (b * Ho + h) * Wo * Co,
-                      in + (b * Hi + h) * Wi * Ci + start * Ci, Wo * Co * elt);
+          std::memcpy(out + ((size_t)b * Ho + h) * Wo * Co,
+                      in + ((size_t)b * Hi + h) * Wi * Ci + (size_t)start * Ci,
+                      (size_t)Wo * Co * elt);
         }
       }
     }
